function for applying enemy patches
